First working version of archive deltas
Taking two archives (only zip flavor has been tested thus far) we can take the
delta, taking into consideration the additions, removals and differences, and
reconstruct a byte-for-byte, checksum equivalent target zip from a source
version and the delta.

This mechanism builds on the normalized entry mechanism of the takari-archiver
where all entries are sorted and the time entries are zeroed out so we can
easily reason about what's in an archive.